Create How to use Github
This link contains various examples on how to use git, including branching, merging, pulling and forking.
I recommend we all read this as GitHub isn't exactly a straight forward thing to understand. We should also 
download git, which is a tool that makes it easy to push or upload your code to your dedicated branches as you code.
Git is something your IDE will use and will make it much easier for us to share our code. Git has several command line tools 
that we need to understand so I hope this article makes it easy to understand.